requirements.txt
google-cloud-bigquery
google-auth
streamlit>=1.22
toml
pandas
db-dtypes